+++ zetaclient/config/types.go
@@ -23,8 +23,10 @@ const (
 	// KeyringBackendFile is the file Cosmos keyring backend
 	KeyringBackendFile KeyringBackend = "file"
 
+	DefaultRelayerDir = "relayer-keys"
+
 	// DefaultRelayerKeyPath is the default path that relayer keys are stored
-	DefaultRelayerKeyPath = "~/.zetacored/relayer-keys"
+	DefaultRelayerKeyPath = "~/.zetacored/" + DefaultRelayerDir
 )
 
 // ClientConfiguration is a subset of zetaclient config that is used by zetacore client
@@ -34,7 +36,6 @@ type ClientConfiguration struct {
 	ChainHomeFolder string `json:"chain_home_folder" mapstructure:"chain_home_folder"`
 	SignerName      string `json:"signer_name"       mapstructure:"signer_name"`
 	SignerPasswd    string `json:"signer_passwd"`
-	HsmMode         bool   `json:"hsm_mode"`
 }
 
 // EVMConfig is the config for EVM chain
@@ -95,8 +96,6 @@ type Config struct {
 	TestTssKeysign      bool           `json:"TestTssKeysign"`
 	KeyringBackend      KeyringBackend `json:"KeyringBackend"`
 	RelayerKeyPath      string         `json:"RelayerKeyPath"`
-	HsmMode             bool           `json:"HsmMode"`
-	HsmHotKey           string         `json:"HsmHotKey"`
 
 	// chain configs
 	EVMChainConfigs map[int64]EVMConfig `json:"EVMChainConfigs"`